[#noissue] Clear up the old real time component and dependency
