+++ web/src/main/webapp/v2/package.json
@@ -36,7 +36,6 @@
     "angular-2-local-storage": "^2.0.0",
     "bowser": "^1.9.4",
     "chart.js": "git+https://github.com/chartjs/Chart.js.git#48fefd92b6dc61345021c6508b23698830ff392f",
-    "chartjs-plugin-streaming": "^1.6.0",
     "core-js": "^2.5.4",
     "gojs": "^1.8.32",
     "js-beautify": "^1.8.8",
