--- web/src/main/webapp/v2/src/app/core/components/real-time-new/new-real-time-agent-chart.component.css
@@ -1,4 +1,0 @@
-:host {
-    display: block;
-    width: 100%;
-}
