--- web/src/main/webapp/v2/src/app/core/components/real-time-new/new-real-time-agent-chart.component.html
@@ -1,8 +1,0 @@
-<pp-new-real-time-chart
-    [activeThreadCounts]="activeThreadCounts"
-    [timeStamp]="timeStamp"
-    [currentPage]="currentPage"
-    [pagingSize]="pagingSize"
-    [chartOption]="chartOption"
-    (outClick)="onClick($event)">
-</pp-new-real-time-chart>
