--- web/src/main/webapp/v2/src/app/core/components/real-time-new/new-real-time-paging-container.component.css
@@ -1,57 +1,0 @@
-:host {
-    display: block;
-    width: 100%;
-    height: 100%;
-}
-.l-thread-chart-wrap {
-    background:#d3dbe6;
-    width:100%;
-    height: 100%;
-}
-.l-agent-chart {
-    width: 100%;
-    max-height: 100%;
-    overflow-y: auto;
-    padding: 5px;
-}
-.l-message {
-    top: 0px;
-    left: 0px;
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    text-align: center;
-    background-color: rgba(226, 226, 226, 0.5);
-    z-index: 9999;
-}
-.l-message h4 {
-    padding: 80px 0 10px 0;
-    font-weight: 100;
-}
-.l-message h4 span {
-    padding: 8px 40px 6px 40px;
-    background-color: #FFF;
-}
-.l-retry button span {
-    margin-right: 6px;
-}
-.l-message.l-no-data h4 span {
-    background-color: #000;
-    color: #FFF;
-}
-.l-paging {
-    width: 100%;
-    background-color: #74879a;
-    color: #FFF;
-    font-size: 15px;
-    margin-bottom: 5px;
-    padding: 11px;
-}
-.l-paging .l-txt {
-    font-weight: lighter;
-}
-.l-paging .l-txt-bold {
-    font-weight: bolder;
-    font-size: 18px;
-    margin-right: 2px;
-}
