--- web/src/main/webapp/v2/src/app/core/components/real-time-new/new-real-time-paging-container.component.html
@@ -1,35 +1,0 @@
-<div class="l-thread-chart-wrap">
-    <div class="l-chart-group-wrap font-opensans">
-        <ng-container *ngIf="activeThreadCounts">
-            <article class="l-agent-chart">
-                <div class="l-paging">
-                    <span class="l-txt">Total Servers : </span>
-                    <span class="l-txt-bold">{{totalCount}}</span>
-                    <span>[ {{firstChartIndex + 1}} ~ {{lastChartIndex + 1}} ]</span>
-                </div>
-                <pp-new-real-time-agent-chart
-                    [timeStamp]="timeStamp"
-                    [activeThreadCounts]="activeThreadCounts"
-                    [currentPage]="currentPage"
-                    [pagingSize]="pagingSize"
-                    (outOpenThreadDump)="onOpenThreadDump($event)"
-                    (outRenderCompleted)="onRenderCompleted()">
-                </pp-new-real-time-agent-chart>
-            </article>
-        </ng-container>
-        <ng-container [ngSwitch]="messageTemplate">
-            <div class="l-message l-waiting" *ngSwitchCase="'LOADING'">
-                <h4><span>Waiting Connection...</span></h4>
-                <i class="fas fa-spinner fa-spin fa-2x fa-fw"></i>
-            </div>
-            <div class="l-message l-retry" *ngSwitchCase="'RETRY'">
-                <h4><span>Closed connection</span></h4>
-                <button type="button" class="pinpoint-btn" (click)="retryConnection()"><span><i class="fas fa-redo"></i></span> Retry</button>
-            </div>            
-            <div class="l-message l-no-data" *ngSwitchCase="'NO_DATA'">
-                <h4><span>This node is not WAS</span></h4>
-            </div>
-            <ng-container *ngSwitchDefault></ng-container>
-        </ng-container>
-    </div>
-</div>
