--- web/src/main/webapp/v2/src/app/core/components/real-time-new/new-real-time-total-chart.component.html
@@ -1,31 +1,0 @@
-<pp-new-real-time-chart
-    [activeThreadCounts]="activeThreadCounts"
-    [timezone]="timezone"
-    [dateFormat]="dateFormat"
-    [timeStamp]="timeStamp"
-    [applicationName]="applicationName"
-    [chartOption]="chartOption"
-    (outSum)="onSum($event)">
-</pp-new-real-time-chart>
-<ul class="l-legend" #legend [ngStyle]="getLegendStyle(legend)">
-    <li>
-        <span class="l-text">Total</span>
-        <span class="l-circle">{{totalCount}}</span>
-    </li>
-    <li>
-        <span class="l-text">Slow</span>
-        <span class="l-circle">{{data[3]}}</span>
-    </li>
-    <li>
-        <span class="l-text">5s</span>
-        <span class="l-circle">{{data[2]}}</span>
-    </li>
-    <li>
-        <span class="l-text">3s</span>
-        <span class="l-circle">{{data[1]}}</span>
-    </li>
-    <li>
-        <span class="l-text">1s</span>
-        <span class="l-circle">{{data[0]}}</span>
-    </li>
-</ul>
