+++ web/src/main/webapp/v2/src/app/core/components/real-time/index.ts
@@ -1,23 +1,23 @@
 import { NgModule } from '@angular/core';
 
 import { SharedModule } from 'app/shared';
-import { RealTimeChartComponent } from './real-time-chart.component';
-import { RealTimeContainerComponent } from './real-time-container.component';
-import { RealTimePagingContainerComponent } from './real-time-paging-container.component';
-import { RealTimeTotalChartComponent } from './real-time-total-chart.component';
-import { RealTimeAgentChartComponent } from './real-time-agent-chart.component';
-import { RealTimeWebSocketService } from './real-time-websocket.service';
-import { ResizeTopDirective } from './resize-top.directive';
 import { HelpViewerPopupModule } from 'app/core/components/help-viewer-popup';
+import { ResizeTopDirective } from 'app/core/components/real-time/resize-top.directive';
+import { RealTimeAgentChartComponent } from 'app/core/components/real-time/real-time-agent-chart.component';
+import { RealTimeChartComponent } from 'app/core/components/real-time/real-time-chart.component';
+import { RealTimeContainerComponent } from 'app/core/components/real-time/real-time-container.component';
+import { RealTimeTotalChartComponent } from 'app/core/components/real-time/real-time-total-chart.component';
+import { RealTimePagingContainerComponent } from 'app/core/components/real-time/real-time-paging-container.component';
+import { RealTimeWebSocketService } from 'app/core/components/real-time/real-time-websocket.service';
 
 @NgModule({
     declarations: [
         ResizeTopDirective,
-        RealTimeChartComponent,
         RealTimeAgentChartComponent,
-        RealTimeTotalChartComponent,
+        RealTimeChartComponent,
         RealTimeContainerComponent,
-        RealTimePagingContainerComponent,
+        RealTimeTotalChartComponent,
+        RealTimePagingContainerComponent
     ],
     imports: [
         SharedModule,
@@ -25,11 +25,7 @@ import { HelpViewerPopupModule } from 'app/core/components/help-viewer-popup';
     ],
     exports: [
         RealTimeContainerComponent,
-        RealTimePagingContainerComponent,
-    ],
-    entryComponents: [
-        RealTimeTotalChartComponent,
-        RealTimeAgentChartComponent,
+        RealTimePagingContainerComponent
     ],
     providers: [
         RealTimeWebSocketService
