+++ web/src/main/webapp/v2/src/app/core/components/real-time/real-time-agent-chart.component.css
@@ -1,53 +1,4 @@
 :host {
     display: block;
-    width: 152px;
-    height: 84px;
-    margin: 0 15px 10px 0px;
-}
-dt {
-    color: #fff;
-    padding: 10px 20px 10px 2px;
-    display: block;
-    overflow: hidden;
-    font-size: 11px;
-    text-align: center;
-    line-height: 1em;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    border-bottom: 1px solid #687b8e;
-    background-color: #74879a;
-}
-i {
-    float: right;
-    cursor: pointer;
-    margin-right: -16px;
-}
-dd {
-    display: flex;
-    flex-flow: row wrap;
-    background: #fff;
-    justify-content: space-around;
-    align-items: center;
-}
-.l-chart-section {
-    background: #fff;
-    border-top: 1px solid #e5e8f0;
-    height: 54px;
     width: 100%;
-    position: relative;
-}
-.l-error-template {
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    left: 0;
-    top: 0;
-    justify-content: center;
-    align-items: center;
-    z-index: 2;
-}
-.l-error-text {
-    font-size: 14px;
-    font-weight: 600;
-    color: #c04e3f;
 }
