+++ web/src/main/webapp/v2/src/app/core/components/real-time/real-time-agent-chart.component.html
@@ -1,15 +1,8 @@
-<dl>
-    <dt><i class="fas fa-external-link-alt" (click)="onOpen()"></i>{{agentName}}</dt>
-    <dd>
-        <div class="l-chart-section">
-            <pp-real-time-chart
-                [style.opacity]="hasError ? 0.5 : 1"
-                [chartData]="chartData"
-                [showAxis]="showAxis">
-            </pp-real-time-chart>
-            <div class="l-error-template" [style.display]="hasError ? 'flex' : 'none'">
-                <p class="l-error-text">{{errorMessage}}</p>
-            </div>
-        </div>
-    </dd>
-</dl>
+<pp-real-time-chart
+    [activeThreadCounts]="activeThreadCounts"
+    [timeStamp]="timeStamp"
+    [currentPage]="currentPage"
+    [pagingSize]="pagingSize"
+    [chartOption]="chartOption"
+    (outClick)="onClick($event)">
+</pp-real-time-chart>
