+++ web/src/main/webapp/v2/src/app/core/components/real-time/real-time-container.component.css
@@ -1,4 +1,5 @@
 :host {
+    display: block;
     z-index: 30;
 }
 .l-thread-chart-wrap {
@@ -45,19 +46,18 @@
     display: flex;
     flex-flow: row wrap;
     padding: 20px 10px 20px 25px;
-    max-height: calc(100% - 34px);
+    height: calc(100% - 34px);
 }
-.l-chart-group-wrap > .l-chart-item {
+.l-chart-group-wrap > .l-total-chart {
     width: 277px;
     height: 178px;
     margin-right: 20px;
 }
-.l-chart-group-wrap .l-chart-group-list {
-    display: flex;
-    flex-flow: row wrap;
-    flex: 1;
+.l-agent-chart {
+    width: calc(100% - 297px);
+    max-height: 100%;
     overflow-y: auto;
-    justify-content: space-between;
+    outline: none;
 }
 .l-message {
     top: 0px;
@@ -89,8 +89,9 @@
     background-color: #74879a;
     color: #FFF;
     font-size: 15px;
-    margin: 5px 5px 4px 5px;
+    margin: 0px 0px 5px 0px;
     padding: 11px;
+    position: relative;
 }
 .l-paging .l-txt {
     font-weight: lighter;
@@ -98,21 +99,45 @@
 .l-paging .l-txt-bold {
     font-weight: bolder;
     font-size: 18px;
+    margin-right: 2px;
 }
-.l-paging .l-page {
-    float: right;
-    cursor: pointer;
-    padding: 0px 6px;
+.fa-chart-area {
+    margin-right: 2px;
 }
-.l-paging .l-page:hover {
-    color: #4b99e3;
+
+.l-paging nav {
+    height: calc(100% - 22px);
+    display: inline-block;
+    position: absolute;
+    right: 10px;
 }
-.l-paging .l-page:last-child {
+
+.l-pagination {
+    display: flex;
+    list-style: none;
+    height: 100%;
+}
+
+.l-page-item {
+    display: block;
+    height: 100%;
+}
+
+.l-page-item:first-child .l-page-link {
     color: #000;
     cursor: initial;
     text-decoration: initial;
     background-color: #FFF;
 }
-.fa-chart-area {
-    margin-right: 2px;
+
+.l-page-link {
+    padding: 0 6px;
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    height: 100%;
+}
+
+.l-page-link:hover {
+    color: #4b99e3;
 }
