+++ web/src/main/webapp/v2/src/app/core/components/real-time/real-time-container.component.html
@@ -4,24 +4,45 @@
             <span class="fas fa-chart-area"></span>Realtime Active Thread Chart
         </div>
         <div class="l-tool-box">
-            <button class="fas fa-thumbtack" [class.l-pin-up]="pinUp" (click)="onPinUp()"></button>
+            <button class="fas fa-thumbtack" [class.l-pin-up]="isPinUp" (click)="onPinUp()"></button>
             <button class="fas fa-question-circle" (click)="onShowHelp($event)"></button>
         </div>
     </div>
     <div class="l-chart-group-wrap font-opensans">
-        <article class="l-chart-item">
-            <ng-container #totalChartPlaceHolder></ng-container>
-        </article>
-        <div class="l-chart-group-list">
-            <div class="l-paging" *ngIf="needPaging()">
-                <span class="l-txt">Total Servers : </span>
-                <span class="l-txt-bold">{{totalCount}}</span>
-                <span>[ 1 ~ {{pagingSize}} ]</span>
-                <span class="l-page" *ngFor="let page of getTotalPage()" (click)="openPage(page)">{{page}}</span>
-                <span class="l-page">1</span>
-            </div>
-            <ng-container #agentChartPlaceHolder></ng-container>
-        </div>
+        <ng-container *ngIf="activeThreadCounts">
+            <article class="l-total-chart">
+                <pp-real-time-total-chart
+                    [timeStamp]="timeStamp"
+                    [timezone]="timezone$ | async"
+                    [dateFormat]="dateFormat$ | async"
+                    [applicationName]="applicationName"
+                    [activeThreadCounts]="activeThreadCounts">
+                </pp-real-time-total-chart>
+            </article>
+            <article class="l-agent-chart">
+                <ng-container *ngIf="needPaging()">
+                    <div class="l-paging">
+                        <span class="l-txt">Total Servers : </span>
+                        <span class="l-txt-bold">{{totalCount}}</span>
+                        <span>[ {{firstChartIndex + 1}} ~ {{lastChartIndex + 1}} ]</span>
+                        <nav>
+                            <ul class="l-pagination">
+                                <li class="l-page-item" *ngFor="let page of getTotalPage()">
+                                    <a class="l-page-link" (click)="openPage(page)">{{page}}</a>
+                                </li>
+                            </ul>
+                        </nav>
+                    </div>
+                </ng-container>
+                <pp-real-time-agent-chart
+                    [timeStamp]="timeStamp"
+                    [activeThreadCounts]="activeThreadCounts"
+                    [pagingSize]="pagingSize"
+                    (outOpenThreadDump)="onOpenThreadDump($event)"
+                    (outRenderCompleted)="onRenderCompleted()">
+                </pp-real-time-agent-chart>
+            </article>
+        </ng-container>
         <ng-container [ngSwitch]="messageTemplate">
             <div class="l-message l-waiting" *ngSwitchCase="'LOADING'">
                 <h4><span>Waiting Connection...</span></h4>
