+++ web/src/main/webapp/v2/src/app/core/components/real-time/real-time-paging-container.component.html
@@ -1,13 +1,22 @@
 <div class="l-thread-chart-wrap">
     <div class="l-chart-group-wrap font-opensans">
-        <div class="l-chart-group-list">
-            <div class="l-paging" *ngIf="needPaging()">
-                <span class="l-txt">Total Servers : </span>
-                <span class="l-txt-bold">{{totalCount}}</span>
-                <span>[ {{startCount + 1}} ~ {{endCount}} ]</span>
-            </div>
-            <ng-container #agentChartPlaceHolder></ng-container>
-        </div>
+        <ng-container *ngIf="activeThreadCounts">
+            <article class="l-agent-chart">
+                <div class="l-paging">
+                    <span class="l-txt">Total Servers : </span>
+                    <span class="l-txt-bold">{{totalCount}}</span>
+                    <span>[ {{firstChartIndex + 1}} ~ {{lastChartIndex + 1}} ]</span>
+                </div>
+                <pp-real-time-agent-chart
+                    [timeStamp]="timeStamp"
+                    [activeThreadCounts]="activeThreadCounts"
+                    [currentPage]="currentPage"
+                    [pagingSize]="pagingSize"
+                    (outOpenThreadDump)="onOpenThreadDump($event)"
+                    (outRenderCompleted)="onRenderCompleted()">
+                </pp-real-time-agent-chart>
+            </article>
+        </ng-container>
         <ng-container [ngSwitch]="messageTemplate">
             <div class="l-message l-waiting" *ngSwitchCase="'LOADING'">
                 <h4><span>Waiting Connection...</span></h4>
