+++ web/src/main/webapp/v2/src/app/core/components/real-time/real-time-total-chart.component.html
@@ -1,39 +1,31 @@
-<dl class="l-wrapper">
-    <dt>{{applicationName}}</dt>
-    <dd>
-        <div class="l-chart-section">
-            <pp-real-time-chart
-                [style.opacity]="hasError ? 0.5 : 1"
-                [chartData]="chartData"
-                [showAxis]="showAxis"
-                [timezone]="timezone"
-                [dateFormat]="dateFormat">
-            </pp-real-time-chart>
-            <div class="l-error-template" [style.display]="hasError ? 'flex' : 'none'">
-                <p class="l-error-text">{{errorMessage}}</p>
-            </div>
-        </div>
-        <ul class="l-legend">
-            <li>
-                <span class="l-text">Total</span>
-                <span class="l-circle">{{getTotalCount()}}</span>
-            </li>
-            <li>
-                <span class="l-text">Slow</span>
-                <span class="l-circle">{{chartData.responseCount[3]}}</span>
-            </li>
-            <li>
-                <span class="l-text">5s</span>
-                <span class="l-circle">{{chartData.responseCount[2]}}</span>
-            </li>
-            <li>
-                <span class="l-text">3s</span>
-                <span class="l-circle">{{chartData.responseCount[1]}}</span>
-            </li>
-            <li>
-                <span class="l-text">1s</span>
-                <span class="l-circle">{{chartData.responseCount[0]}}</span>
-            </li>
-        </ul>
-    </dd>
-</dl>
+<pp-real-time-chart
+    [activeThreadCounts]="activeThreadCounts"
+    [timezone]="timezone"
+    [dateFormat]="dateFormat"
+    [timeStamp]="timeStamp"
+    [applicationName]="applicationName"
+    [chartOption]="chartOption"
+    (outSum)="onSum($event)">
+</pp-real-time-chart>
+<ul class="l-legend" #legend [ngStyle]="getLegendStyle(legend)">
+    <li>
+        <span class="l-text">Total</span>
+        <span class="l-circle">{{totalCount}}</span>
+    </li>
+    <li>
+        <span class="l-text">Slow</span>
+        <span class="l-circle">{{data[3]}}</span>
+    </li>
+    <li>
+        <span class="l-text">5s</span>
+        <span class="l-circle">{{data[2]}}</span>
+    </li>
+    <li>
+        <span class="l-text">3s</span>
+        <span class="l-circle">{{data[1]}}</span>
+    </li>
+    <li>
+        <span class="l-text">1s</span>
+        <span class="l-circle">{{data[0]}}</span>
+    </li>
+</ul>
