+++ web/src/main/webapp/v2/src/app/routes/main-page/index.ts
@@ -9,7 +9,6 @@ import { PeriodSelectorModule } from 'app/core/components/period-selector';
 import { CommandGroupModule } from 'app/core/components/command-group';
 import { MainContentsModule } from 'app/core/components/main-contents';
 import { RealTimeModule } from 'app/core/components/real-time';
-import { NewRealTimeModule } from 'app/core/components/real-time-new';
 import { SideBarModule } from 'app/core/components/side-bar';
 import { MainPageComponent } from './main-page.component';
 import { routing } from './main-page.routing';
@@ -28,14 +27,11 @@ import { HelpViewerPopupModule } from 'app/core/components/help-viewer-popup';
         CommandGroupModule,
         MainContentsModule,
         RealTimeModule,
-        NewRealTimeModule,
         SideBarModule,
         HelpViewerPopupModule,
         RouterModule.forChild(routing)
     ],
-    exports: [
-
-    ],
+    exports: [],
     providers: []
 })
 export class MainPageModule { }
