+++ web/src/main/webapp/v2/src/app/routes/main-page/main-page.component.html
@@ -14,7 +14,7 @@
         <article class="l-main-contents">
             <router-outlet></router-outlet>
         </article>
-        <pp-new-real-time-container *ngIf="enableRealTime$ | async"></pp-new-real-time-container>
+        <pp-real-time-container *ngIf="enableRealTime$ | async"></pp-real-time-container>
     </section>
     <pp-side-bar-container></pp-side-bar-container>
 </div>
