+++ web/src/main/webapp/v2/src/app/routes/main-page/main-page.routing.ts
@@ -6,9 +6,7 @@ import { EmptyContentsComponent, NoneComponent } from 'app/shared/components/emp
 import { UrlRedirectorComponent } from 'app/shared/components/url-redirector';
 // import { UrlValidateGuard } from 'app/shared/services';
 import { SystemConfigurationResolverService, ApplicationListResolverService, ServerTimeResolverService } from 'app/shared/services';
-
 import { MainPageComponent } from './main-page.component';
-import { NewRealTimeContainerComponent } from 'app/core/components/real-time-new/new-real-time-container.component';
 
 export const routing: Routes = [
     {
